+++ src/Pages/Studies/EditStudy.tsx
@@ -1,0 +1,9 @@
+import { useEditStudyMutation } from '@/Hooks/study/useEditStudy';
+import Layout from './Layout';
+import { useParams } from 'react-router-dom';
+
+export const EditStudy = () => {
+  const { studyId } = useParams();
+
+  return <Layout {...useEditStudyMutation(parseInt(studyId))} />;
+};
